// src/app/modules/products/television/television.interface.ts
export interface ITelevisionSpecifications {
  color: string;
  screenSize: string;
  resolution: string;
  viewingAngle: string;
  soundSystem: string;
  source: string[];
  supportedVideoFormat: string[];
  bodyDesign: string;
  banglaSupportLanguage: string;
  backlightAdjustable: string;
  pictureMode: string[];
  powerConsumption: string;
  tvSystem: string[];
  backlightType: string;
  colorDepth: string;
  aspectRatio: string[];
  thunderProtective: string;
  remote: string;
  wallMountBracket: string;
  contrastRatio: string;
  brightness: string;
  speaker: string;
  channelProgram: string;
  warranty: string;
}